Using int type everywhere instaed of int/int64

+++ objects/nodeStatstics.go
@@ -2,7 +2,7 @@ package objects
 
 // NodeStatistics is the structure for node statistics
 type NodeStatistics struct {
-	ID                int64   `json:"id"`
+	ID                int   `json:"id"`
 	Type              string  `json:"type"`
 	Name              string  `json:"name"`
 	Status            string  `json:"status"`
@@ -10,8 +10,8 @@ type NodeStatistics struct {
 	StartTime         float64 `json:"startTime"`
 	EndTime           float64 `json:"endTime"`
 	Duration          float64 `json:"duration"`
-	DatasetRows       int64   `json:"datasetRows"`
-	DatasetCols       int64   `json:"datasetCols"`
+	DatasetRows       int   `json:"datasetRows"`
+	DatasetCols       int   `json:"datasetCols"`
 	FreeMemoryInitial float64 `json:"freeMemoryInitial"`
 	FreeMemoryFinal   float64 `json:"freeMemoryFinal"`
 	FreeDiscInitial   float64 `json:"freeDiscInitial"`
